Move and rename env_role core policies to core/environments module
