+++ backend/dataall/core/environment/cdk/env_role_core_policies/__init__.py
@@ -1,6 +1,6 @@
 """Contains the code for creating environment policies"""
 
-from dataall.base.cdkproxy.stacks.policies import (
+from dataall.core.environment.cdk.env_role_core_policies import (
     cloudformation, data_policy, service_policy, athena, secretsmanager, sqs, ssm
 )
