+++ backend/dataall/core/environment/cdk/env_role_core_policies/secretsmanager.py
